pyproject.toml
[tool.poetry]
name = "yt-sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["Maja <majam1616@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
youtube-transcript-api = "^0.4.4"
boto3 = "^1.22.1"
dynamodb-json = "^1.3"
moto = {extras = ["dynamodb"], version = "^3.1.7"}

[tool.poetry.dev-dependencies]
ipykernel = "^6.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
